Fix compile

+++ src/ArchivItem_Sound.cpp
@@ -95,7 +95,8 @@ libsiedler2::baseArchivItem_Sound* libsiedler2::baseArchivItem_Sound::findSubTyp
         // midi
         item = dynamic_cast<baseArchivItem_Sound*>(getAllocator().create(BOBTYPE_SOUND, SOUNDTYPE_MIDI));
     }
-    else if(strncmp(header, "OggS", 4) == 0 || strncmp(header, "ID3", 3) == 0 || (header[0] == 0xFF && header[1] == 0xFB) )
+    else if(strncmp(header, "OggS", 4) == 0 || strncmp(header, "ID3", 3) == 0 ||
+        (static_cast<uint8_t>(header[0]) == 0xFF && static_cast<uint8_t>(header[1]) == 0xFB) )
     {
         // ogg, mp3 (id3tag, ohne),
         item = dynamic_cast<baseArchivItem_Sound*>(getAllocator().create(BOBTYPE_SOUND, SOUNDTYPE_OTHER));
